trying to see if removing the route and hardcoding the page solves th… (#25)
## Motivation

<!--
Briefly describe the goal(s) of this PR.
-->
Fix the deployment bug where the page would be blank

## Proposal

<!--
Summarize the proposed changes and how they address the goal(s) stated
above.
-->
Because we don't have multiple pages we can hardcode the page for now so
that `https://apps.linera.net/gol/index.html` does not have problems
serving the page because of the additional `gol/index.html` params

## Test Plan

<!--
Explain how you made sure that the changes are correct and that they
perform as intended.
-->
Lele deployed it and it works: `https://apps.linera.net/gol/index.html`