fix resetToInitialState bug (#26)
## Motivation

<!--
Briefly describe the goal(s) of this PR.
-->
fixed bug where the reset to initial state would not clear the history 

## Proposal

<!--
Summarize the proposed changes and how they address the goal(s) stated
above.
-->
when reseting the board to the initial state, clear the history as well 

## Test Plan

<!--
Explain how you made sure that the changes are correct and that they
perform as intended.
-->
tested it locally