Multiple scoring chains (Frontend) (#40)
## Motivation

Update the frontend to finish the work after #39

## Proposal

* Update the APP_ID and other constants to follow the new design
* Make the new app query the puzzles solved by the previous one
* Use 4 scoring chains

The description of the scoring chains and the mapping from user address
is in the web app itself. This means we can change it later.

CAVEAT: By allowing several scoring chains that can change, we're making
the tracking of scores in the portal more difficult in several ways.
* Old scoring chains must always be queried for some time because users
may still use the old app.
* When we change the set of chains, reports of solved puzzles may become
duplicated.

## Test Plan

Manually tested with test user (old app then new app)